+++ src/OmniSharp.Abstractions/Models/v1/GotoDefinition/GotoDefinitionRequest.cs
@@ -5,7 +5,7 @@ namespace OmniSharp.Models.GotoDefinition
     [OmniSharpEndpoint(OmniSharpEndpoints.GotoDefinition, typeof(GotoDefinitionRequest), typeof(GotoDefinitionResponse))]
     public class GotoDefinitionRequest : Request
     {
-        public int Timeout { get; set; } = 2000;
+        public int Timeout { get; set; } = 10000;
         public bool WantMetadata { get; set; }
     }
 }