+++ tests/OmniSharp.Roslyn.CSharp.Tests/SemanticHighlightFacts.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Threading.Tasks;
-using System.Windows.Input;
 using Microsoft.CodeAnalysis.Text;
 using OmniSharp.Models.SemanticHighlight;
 using OmniSharp.Models.V2;
@@ -19,6 +18,22 @@ namespace OmniSharp.Roslyn.CSharp.Tests
         }
 
         protected override string EndpointName => OmniSharpEndpoints.V2.Highlight;
+
+        [Fact]
+        public async Task InvalidPositionDoesNotThrow()
+        {
+            var testFile = new TestFile("a.cs", @"
+namespace N1
+{
+    class C1 { int n = true; }
+}
+");
+
+            var line = -1;
+            var highlights = await GetSemanticHighlightsForLineAsync(testFile, line);
+
+            Assert.Empty(highlights);
+        }
 
         [Fact]
         public async Task SemanticHighlightSingleLine()
@@ -222,6 +237,44 @@ static class C1
             );
         }
 
+        [Fact]
+        public async Task SemanticHighlightRecordName()
+        {
+            var testFile = new TestFile("a.cs", @"
+R1 r1 = new R1(string.Empty, 1);
+record R1(string S, int I);
+");
+
+            var highlights = await GetSemanticHighlightsForFileAsync(testFile);
+
+            AssertSyntax(highlights, testFile.Content.Code, 0,
+                ClassName("R1"),
+                Local("r1"),
+                Operator("="),
+                Keyword("new"),
+                ClassName("R1"),
+                Punctuation("("),
+                Keyword("string"),
+                Operator("."),
+                Field("Empty", SemanticHighlightModifier.Static),
+                Punctuation(","),
+                Number("1"),
+                Punctuation(")"),
+                Punctuation(";"),
+
+                Keyword("record"),
+                ClassName("R1"),
+                Punctuation("("),
+                Keyword("string"),
+                Parameter("S"),
+                Punctuation(","),
+                Keyword("int"),
+                Parameter("I"),
+                Punctuation(")"),
+                Punctuation(";")
+            );
+        }
+
         private Task<SemanticHighlightSpan[]> GetSemanticHighlightsForFileAsync(TestFile testFile)
         {
             return GetSemanticHighlightsAsync(testFile, range: null);
@@ -301,6 +354,7 @@ static class C1
         private static (SemanticHighlightClassification type, string text, SemanticHighlightModifier[] modifiers) ClassName(string text, params SemanticHighlightModifier[] modifiers) => (SemanticHighlightClassification.ClassName, text, modifiers);
         private static (SemanticHighlightClassification type, string text, SemanticHighlightModifier[] modifiers) Field(string text, params SemanticHighlightModifier[] modifiers) => (SemanticHighlightClassification.FieldName, text, modifiers);
         private static (SemanticHighlightClassification type, string text, SemanticHighlightModifier[] modifiers) Identifier(string text, params SemanticHighlightModifier[] modifiers) => (SemanticHighlightClassification.Identifier, text, modifiers);
+        private static (SemanticHighlightClassification type, string text, SemanticHighlightModifier[] modifiers) Parameter(string text, params SemanticHighlightModifier[] modifiers) => (SemanticHighlightClassification.ParameterName, text, modifiers);
         private static (SemanticHighlightClassification type, string text, SemanticHighlightModifier[] modifiers) NamespaceName(string text, params SemanticHighlightModifier[] modifiers) => (SemanticHighlightClassification.NamespaceName, text, modifiers);
         private static (SemanticHighlightClassification type, string text, SemanticHighlightModifier[] modifiers) Keyword(string text, params SemanticHighlightModifier[] modifiers) => (SemanticHighlightClassification.Keyword, text, modifiers);
         private static (SemanticHighlightClassification type, string text, SemanticHighlightModifier[] modifiers) ControlKeyword(string text, params SemanticHighlightModifier[] modifiers) => (SemanticHighlightClassification.ControlKeyword, text, modifiers);